Adding vaccination endpoints

+++ src/main/api/drugs/business.py
@@ -1,10 +1,10 @@
 from http import HTTPStatus
 
 from flask_restx import abort
-from flask import jsonify
 
 from main.api.token.decorators import token_required
 from main.models.drug import Drug
+from main.models.vaccination import Vaccination
 
 
 @token_required
@@ -13,10 +13,9 @@ def get_drug(id):
         drug = Drug.query.get(id)
     except Exception:
         abort(HTTPStatus.INTERNAL_SERVER_ERROR)
-    if drug:
-        return drug
-    else:
+    if not drug:
         abort(HTTPStatus.NOT_FOUND)
+    return drug
 
 
 @token_required
@@ -30,11 +29,11 @@ def list_drugs():
 
 @token_required
 def create_drug(drug):
-    d = Drug(name=drug["name"], code=drug["code"], description=drug["description"])
+    d = Drug(**drug)
     if d.save():
         return d, HTTPStatus.CREATED
     else:
-        abort(HTTPStatus.INTERNAL_SERVER_ERROR, "Drug code already exists.")
+        abort(HTTPStatus.INTERNAL_SERVER_ERROR, "Drug code already exists")
 
 
 @token_required
@@ -48,7 +47,7 @@ def update_drug(drug, id):
     if d.save():
         return True, HTTPStatus.OK
     else:
-        abort(HTTPStatus.INTERNAL_SERVER_ERROR, "Drug code already exists.")
+        abort(HTTPStatus.INTERNAL_SERVER_ERROR, "Drug code already exists")
 
 
 @token_required
@@ -56,7 +55,10 @@ def delete_drug(id):
     d = Drug.query.get(id)
     if not d:
         abort(HTTPStatus.NOT_FOUND)
+    vac = Vaccination.query.filter_by(drug=id)
+    if vac:
+        abort(HTTPStatus.BAD_REQUEST, "Drug has vaccinations scheduled")
     if d.delete():
         return True, HTTPStatus.NO_CONTENT
     else:
-        abort(HTTPStatus.INTERNAL_SERVER_ERROR, "Error while deleting drug record.")
+        abort(HTTPStatus.INTERNAL_SERVER_ERROR, "Error while deleting drug record")
